+++ Managerment/Final/src/main/java/com/example/tien/Final/repos/EmployeeRepository.java
@@ -4,6 +4,6 @@ import com.example.tien.Final.entity.Employee;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 
 }
