+++ Managerment/Final/src/main/java/com/example/tien/Final/repos/PositionRepository.java
@@ -3,7 +3,7 @@ package com.example.tien.Final.repos;
 import com.example.tien.Final.entity.Position;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PositionRepository extends JpaRepository<Position, Integer> {
+public interface PositionRepository extends JpaRepository<Position, Long> {
 
 
     Position findAllByName(String name);
